use AutoSubmitFilterForm in FilterForm
