+++ packages/ra-ui-materialui/src/list/filter/FilterForm.spec.tsx
@@ -20,7 +20,7 @@ import {
     WithAutoCompleteArrayInput,
     WithComplexValueFilter,
 } from './FilterButton.stories';
-import { FilterForm, getFilterFormValues } from './FilterForm';
+import { FilterForm } from './FilterForm';
 
 describe('<FilterForm />', () => {
     const defaultListContext = {
@@ -125,11 +125,7 @@ describe('<FilterForm />', () => {
             target: { value: 'foo' },
         });
         await waitFor(() => {
-            expect(setFilters).toHaveBeenCalledWith(
-                { title: 'foo' },
-                { title: true },
-                true
-            );
+            expect(setFilters).toHaveBeenCalledWith({ title: 'foo' });
         });
     });
 
@@ -293,38 +289,6 @@ describe('<FilterForm />', () => {
         });
 
         await screen.findByText('1-2 of 2');
-    });
-
-    describe('getFilterFormValues', () => {
-        it('should correctly get the filter form values from the new filterValues', () => {
-            const currentFormValues = {
-                classicToClear: 'abc',
-                nestedToClear: { nestedValue: 'def' },
-                classicUpdated: 'ghi',
-                nestedUpdated: { nestedValue: 'jkl' },
-                nestedToSet: { nestedValue: undefined },
-                published_at: new Date('2022-01-01T03:00:00.000Z'),
-                clearedDateValue: null,
-            };
-            const newFilterValues = {
-                classicUpdated: 'ghi2',
-                nestedUpdated: { nestedValue: 'jkl2' },
-                nestedToSet: { nestedValue: 'mno2' },
-                published_at: '2022-01-01T03:00:00.000Z',
-            };
-
-            expect(
-                getFilterFormValues(currentFormValues, newFilterValues)
-            ).toEqual({
-                classicToClear: '',
-                nestedToClear: { nestedValue: '' },
-                classicUpdated: 'ghi2',
-                nestedUpdated: { nestedValue: 'jkl2' },
-                nestedToSet: { nestedValue: 'mno2' },
-                published_at: '2022-01-01T03:00:00.000Z',
-                clearedDateValue: '',
-            });
-        });
     });
 
     it('should not reapply previous filter form values when clearing nested AutocompleteArrayInput', async () => {
